Refactoring the Sink Registry
The initial design of `gominion` had all the modules registered
automatically to start working when the broker starts.

However, the Sink API is distinct compared to the RPC API. For Sink,
there is a state, as each implementation needs to track UDP/TCP
listeners, among other information. In contrast, the RPC modules
are stateless, you only need to run a method, and there is nothing
kept outside the context of its execution.

The purpose of this change is to run multiple brokers simultaneously,
all with their own listeners, to be able to perform scale tests.
Although, this won't introduce any behavior change when running a
single standalone `gominion`.